pyproject.toml
[project]
name = "cbam-resnet-waste-classification"
version = "0.1.0"
description = "Industry-grade waste classification using CBAM-ResNet34 from scratch"
authors = [{name = "Your Name"}]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "pillow>=9.0.0",
    "tqdm>=4.64.0",
    "scikit-learn>=1.0.0",
    "flask>=2.2.0",
    "onnx>=1.12.0",
    "onnxruntime>=1.12.0",
    "opencv-python>=4.7.0",
    "requests>=2.28.0",
    "pandas>=1.5.0",
    "seaborn>=0.11.0",
    "kaggle>=1.5.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"